Cleanup graph dataset iterators
